requirements.txt
langchain
langchain_openai
langchain-community 
pymupdf

python-dotenv
streamlit 